fix: move formatSuggestionItem to separated file
