+++ src/components/GAddressLookup/component.js
@@ -9,33 +9,8 @@ import SelectedLocationIcon from './icons/selectedLocationIcon';
 import SearchValueIcon from './icons/searchValueIcon';
 import PoweredByGoogleLogo from './icons/poweredByGoogle';
 import { uniqueId } from '../../libs/utils';
+import { formatSuggestionItem } from './functions';
 import './styles.css';
-
-function formatSuggestionItem(suggestion, highlightMatch, icon) {
-    let formattedLabel;
-
-    if (highlightMatch) {
-        formattedLabel = suggestion.structured_formatting.main_text_matched_substrings.reduceRight(
-            (previousText, currentMatch) => {
-                const matchedTerm = previousText.slice(currentMatch.offset, currentMatch.length);
-
-                return suggestion.structured_formatting.main_text.replace(
-                    matchedTerm,
-                    `<b>${matchedTerm}</b>`,
-                );
-            },
-            suggestion.structured_formatting.main_text,
-        );
-    } else {
-        formattedLabel = suggestion.description;
-    }
-
-    return {
-        label: formattedLabel,
-        description: suggestion.structured_formatting.secondary_text,
-        icon,
-    };
-}
 
 class PlacesLookupComponent extends Component {
     constructor(props) {
